--- src/ecosolutions/CustomerMain/controller/deliverypoint/dpmainController.java
@@ -1,41 +1,0 @@
-package ecosolutions.CustomerMain.controller.deliverypoint;
-
-import com.jfoenix.controls.JFXButton;
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-
-public class dpmainController {
-
-    @FXML
-    private JFXButton btncustomer;
-
-    @FXML
-    private JFXButton btnconfirm;
-
-    @FXML
-    private JFXButton btnverify;
-
-    @FXML
-    private JFXButton btnlogout;
-
-    @FXML
-    void handleConfirmation(ActionEvent event) {
-
-    }
-
-    @FXML
-    void handleCustomer(ActionEvent event) {
-
-    }
-
-    @FXML
-    void handleLogOut(ActionEvent event) {
-
-    }
-
-    @FXML
-    void handleVerification(ActionEvent event) {
-
-    }
-
-}
