+++ src/ecosolutions/presentation/views/DeliveryPointView.fxml
@@ -42,33 +42,33 @@
                               <Font name="Myanmar Text" size="15.0" />
                            </font>
                         </Label>
-                        <TextField fx:id="qtyTextField" layoutX="221.0" layoutY="149.0" prefHeight="37.0" prefWidth="130.0" />
-                        <Label alignment="CENTER" layoutX="221.0" layoutY="105.0" prefHeight="37.0" prefWidth="130.0" style="-fx-background-color: #424242;" text="Insert Quantity" textFill="WHITE">
+                        <TextField fx:id="qtyTextField" layoutX="215.0" layoutY="149.0" prefHeight="37.0" prefWidth="130.0" />
+                        <Label alignment="CENTER" layoutX="215.0" layoutY="105.0" prefHeight="37.0" prefWidth="130.0" style="-fx-background-color: #424242;" text="Insert Quantity" textFill="WHITE">
                            <font>
                               <Font name="Myanmar Text" size="15.0" />
                            </font>
                         </Label>
-                        <JFXButton fx:id="btnEdit" layoutX="230.0" layoutY="258.0" onAction="#handleEditOrder" prefHeight="37.0" prefWidth="111.0" text="Edit" textFill="WHITE">
+                        <JFXButton fx:id="btnEdit" layoutX="215.0" layoutY="259.0" onAction="#handleEditOrder" prefHeight="41.0" prefWidth="134.0" text="Edit Order" textFill="WHITE">
                            <font>
                               <Font name="Myanmar Text" size="12.0" />
                            </font>
                         </JFXButton>
-                        <JFXButton fx:id="btnAdd" layoutX="231.0" layoutY="200.0" onAction="#addToBasket" prefHeight="37.0" prefWidth="111.0" text="Add" textFill="WHITE">
+                        <JFXButton fx:id="btnAdd" layoutX="215.0" layoutY="206.0" onAction="#addToBasket" prefHeight="37.0" prefWidth="134.0" text="Add Item" textFill="WHITE">
                            <font>
                               <Font name="Myanmar Text" size="12.0" />
                            </font>
                         </JFXButton>
-                        <JFXButton fx:id="btnreturn" layoutX="231.0" layoutY="314.0" onAction="#handleReturn" prefHeight="37.0" prefWidth="111.0" text="Register" textFill="WHITE">
+                        <JFXButton fx:id="btnreturn" layoutX="215.0" layoutY="314.0" onAction="#handleReturn" prefHeight="41.0" prefWidth="134.0" text="Return to Main" textFill="WHITE">
                            <font>
                               <Font name="Myanmar Text" size="12.0" />
                            </font>
                         </JFXButton>
-                        <Label alignment="CENTER" layoutX="221.0" layoutY="15.0" prefHeight="37.0" prefWidth="130.0" style="-fx-background-color: #424242;" text="Insert CustomerID" textFill="WHITE">
+                        <Label alignment="CENTER" layoutX="215.0" layoutY="15.0" prefHeight="37.0" prefWidth="130.0" style="-fx-background-color: #424242;" text="Insert CustomerID" textFill="WHITE">
                            <font>
                               <Font name="Myanmar Text" size="15.0" />
                            </font>
                         </Label>
-                        <TextField fx:id="cusidTextField" layoutX="221.0" layoutY="57.0" prefHeight="38.0" prefWidth="130.0" />
+                        <TextField fx:id="cusidTextField" layoutX="215.0" layoutY="57.0" prefHeight="38.0" prefWidth="130.0" />
                      </children>
                   </Pane>
                </center>
