+++ src/ecosolutions/presentation/views/deliverypointmain.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.text.Font?>
 
 
-<AnchorPane prefHeight="326.0" prefWidth="264.0" stylesheets="@../../../../resources/StyleSheets/color-style.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ecosolutions.CustomerMain.controller.deliverypoint.dpmainController">
+<AnchorPane prefHeight="326.0" prefWidth="264.0" stylesheets="@../../../resources/StyleSheets/color-style.css" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ecosolutions.presentation.controllers.dpmainController">
    <children>
       <JFXButton fx:id="btncustomer" layoutX="41.0" layoutY="59.0" onAction="#handleCustomer" prefHeight="41.0" prefWidth="182.0" text="New Customer" />
       <JFXButton fx:id="btnconfirm" layoutX="41.0" layoutY="115.0" onAction="#handleConfirmation" prefHeight="41.0" prefWidth="182.0" text="Confirm Order" />
